refactor: removido a função validateParams() do arquivo createUserAccount.js para o aquivo utils.js na pasta src\services
